+++ rasa/cli/test.py
@@ -4,7 +4,6 @@ import os
 from typing import List
 
 from rasa.cli.arguments import test as arguments
-from rasa.cli.utils import get_validated_path
 from rasa.constants import (
     DEFAULT_CONFIG_PATH,
     DEFAULT_DATA_PATH,
@@ -14,8 +13,9 @@ from rasa.constants import (
     DEFAULT_NLU_RESULTS_PATH,
     CONFIG_SCHEMA_FILE,
 )
-from rasa.test import test_core_models, compare_nlu_models, test_core_models_in_dir
-from rasa.utils.validation import validate_yaml_schema, InvalidYamlFileError
+import rasa.test as rasa_test
+import rasa.utils.validation as validation_utils
+import rasa.cli.utils as cli_utils
 
 logger = logging.getLogger(__name__)
 
@@ -61,10 +61,10 @@ def test_core(args: argparse.Namespace) -> None:
     from rasa import data
     from rasa.test import test_core
 
-    endpoints = get_validated_path(
+    endpoints = cli_utils.get_validated_path(
         args.endpoints, "endpoints", DEFAULT_ENDPOINTS_PATH, True
     )
-    stories = get_validated_path(args.stories, "stories", DEFAULT_DATA_PATH)
+    stories = cli_utils.get_validated_path(args.stories, "stories", DEFAULT_DATA_PATH)
     stories = data.get_core_directory(stories)
     output = args.out or DEFAULT_RESULTS_PATH
 
@@ -75,10 +75,12 @@ def test_core(args: argparse.Namespace) -> None:
         args.model = args.model[0]
 
     if isinstance(args.model, str):
-        model_path = get_validated_path(args.model, "model", DEFAULT_MODELS_PATH)
+        model_path = cli_utils.get_validated_path(
+            args.model, "model", DEFAULT_MODELS_PATH
+        )
 
         if args.evaluate_models_in_dir:
-            test_core_models_in_dir(args.model, stories, output)
+            rasa_test.test_core_models_in_directory(args.model, stories, output)
         else:
             test_core(
                 model=model_path,
@@ -89,15 +91,14 @@ def test_core(args: argparse.Namespace) -> None:
             )
 
     else:
-        test_core_models(args.model, stories, output)
+        rasa_test.test_core_models(args.model, stories, output)
 
 
 def test_nlu(args: argparse.Namespace) -> None:
     from rasa import data
-    from rasa.test import test_nlu, perform_nlu_cross_validation
-    import rasa.utils.io
+    import rasa.utils.io as io_utils
 
-    nlu_data = get_validated_path(args.nlu, "nlu", DEFAULT_DATA_PATH)
+    nlu_data = cli_utils.get_validated_path(args.nlu, "nlu", DEFAULT_DATA_PATH)
     nlu_data = data.get_nlu_directory(nlu_data)
 
     if args.config is not None and len(args.config) == 1:
@@ -118,20 +119,20 @@ def test_nlu(args: argparse.Namespace) -> None:
         config_files = []
         for file in args.config:
             try:
-                validate_yaml_schema(
-                    rasa.utils.io.read_file(file),
+                validation_utils.validate_yaml_schema(
+                    io_utils.read_file(file),
                     CONFIG_SCHEMA_FILE,
                     show_validation_errors=False,
                 )
                 config_files.append(file)
-            except InvalidYamlFileError:
+            except validation_utils.InvalidYamlFileError:
                 logger.debug(
                     "Ignoring file '{}' as it is not a valid config file.".format(file)
                 )
                 continue
 
         output = args.report or DEFAULT_NLU_RESULTS_PATH
-        compare_nlu_models(
+        rasa_test.compare_nlu_models(
             configs=config_files,
             nlu=nlu_data,
             output=output,
@@ -140,11 +141,15 @@ def test_nlu(args: argparse.Namespace) -> None:
         )
     elif args.cross_validation:
         logger.info("Test model using cross validation.")
-        config = get_validated_path(args.config, "config", DEFAULT_CONFIG_PATH)
-        perform_nlu_cross_validation(config, nlu_data, vars(args))
+        config = cli_utils.get_validated_path(
+            args.config, "config", DEFAULT_CONFIG_PATH
+        )
+        rasa_test.perform_nlu_cross_validation(config, nlu_data, vars(args))
     else:
-        model_path = get_validated_path(args.model, "model", DEFAULT_MODELS_PATH)
-        test_nlu(model_path, nlu_data, vars(args))
+        model_path = cli_utils.get_validated_path(
+            args.model, "model", DEFAULT_MODELS_PATH
+        )
+        rasa_test.test_nlu(model_path, nlu_data, vars(args))
 
 
 def test(args: argparse.Namespace):